+++ kvdb/CHANGELOG.md
@@ -5,7 +5,15 @@ The format is based on [Keep a Changelog].
 [Keep a Changelog]: http://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
+### Breaking
+- Removed `write_buffered` and `flush` methods. [#313](https://github.com/paritytech/parity-common/pull/313)
+- Introduce a new `DeletePrefix` database operation. [#360](https://github.com/paritytech/parity-common/pull/360)
+- Rename prefix iteration to `iter_with_prefix`. [#365](https://github.com/paritytech/parity-common/pull/365)
+
+## [0.5.0] - 2020-03-16
 - License changed from GPL3 to dual MIT/Apache2. [#342](https://github.com/paritytech/parity-common/pull/342)
+- Remove dependency on parity-bytes. [#351](https://github.com/paritytech/parity-common/pull/351)
+- Updated dependencies. [#361](https://github.com/paritytech/parity-common/pull/361)
 
 ## [0.4.0] - 2019-01-06
 - Bump parking_lot to 0.10. [#332](https://github.com/paritytech/parity-common/pull/332)